requirements.txt
# Prompt Graveyard Dependencies 🎃
# Core evaluation and LLM integration

# Async HTTP client for LLM API calls
aiohttp>=3.8.0

# YAML parsing for configuration
PyYAML>=6.0

# Vector embeddings and similarity search
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Data processing and statistics
numpy>=1.24.0
pandas>=2.0.0

# Optional: For more advanced NLP metrics
# nltk>=3.8
# spacy>=3.6.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0